keystore: use consistent types for seed length
Before this commit, we used size_t, uint32_t and uint8_t for seed
sizes, which can only be 16, 24 or 32 bytes long.

uint8_t is only needed when storing the encrypted seed length to flash
memory (see `memory_set_encrypted_seed_and_hmac()`) and when reading
it back. We only keep uint8_t there and use size_t everywhere else for
seed lenghts.

+++ test/unit-test/test_keystore_functional.c
@@ -64,7 +64,7 @@ static void _test_seeds(void** state)
     _smarteeprom_reset();
     assert_true(keystore_is_locked());
     uint8_t read_seed[KEYSTORE_MAX_SEED_LENGTH];
-    uint32_t read_seed_len;
+    size_t read_seed_len;
     assert_false(keystore_copy_seed(read_seed, &read_seed_len));
 
     will_return(__wrap_memory_is_initialized, true);
